refactor(translator): merge auto-translate core into core/domain/auto-translate
D1: fold core/auto-translate (drift predicate) + core/auto-translate-config (types,
custom key, reader) into one core/domain/auto-translate/ with a single barrel —
mirroring the provenance interface+predicate colocation. Kills the 4-places/3-near-
identical-names friction (4 -> 3, each in a distinct layer). Pure moves; importers
updated; public API unchanged.
