+++ packages/payload-plugin-translator/src/client/entities/translation/model/autoTranslateSummary.ts
@@ -1,4 +1,4 @@
-import { getAutoTranslateConfig } from "../../../../core/auto-translate-config";
+import { getAutoTranslateConfig } from "../../../../core/domain/auto-translate";
 
 /** The auto-translate facts the marker renders — resolved from a collection's opt-in config. */
 export type AutoTranslateSummary = { targets: string[]; sourceLocale: string };
