--- packages/payload-plugin-translator/src/core/auto-translate/index.ts
@@ -1,1 +1,0 @@
-export { hasSourceContentChanged } from "./hasSourceContentChanged";
