+++ packages/payload-plugin-translator/src/core/domain/auto-translate/getAutoTranslateConfig.test.ts
