+++ packages/payload-plugin-translator/src/core/domain/auto-translate/getAutoTranslateConfig.ts
@@ -1,4 +1,4 @@
-import { isObject } from "../kernel/utils/isObject";
+import { isObject } from "../../kernel/utils/isObject";
 
 import type { AutoTranslateConfig } from "./types";
 import { AUTO_TRANSLATE_CUSTOM_KEY } from "./types";
