+++ packages/payload-plugin-translator/src/core/domain/auto-translate/hasSourceContentChanged.ts
@@ -1,5 +1,5 @@
-import { computeSourceFingerprint } from "../domain/content-projection/computeSourceFingerprint";
-import type { FieldLike } from "../kernel/field-traversal";
+import { computeSourceFingerprint } from "../content-projection/computeSourceFingerprint";
+import type { FieldLike } from "../../kernel/field-traversal";
 
 /**
  * Whether a source-locale save actually changed translatable content — the auto-translate drift-gate,
