+++ packages/payload-plugin-translator/src/core/domain/auto-translate/index.ts
@@ -1,3 +1,4 @@
 export { getAutoTranslateConfig } from "./getAutoTranslateConfig";
+export { hasSourceContentChanged } from "./hasSourceContentChanged";
 export { AUTO_TRANSLATE_CUSTOM_KEY } from "./types";
 export type { AutoTranslateConfig, AutoTranslateStrategy } from "./types";
